Title: 최솟값 3개, Time: 375ms, Memory: 23MB, Status: Wrong Answer - CodeTree

+++ 240810/최솟값 3개/top-3-smallest-number.java
@@ -11,7 +11,7 @@ public class Main {
         int n = Integer.parseInt(br.readLine());
         PriorityQueue<Integer> pq = new PriorityQueue<>((a,b) -> b - a);
 
-        int total = 1;
+        long total = 1;
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i < n; i++) {
             int num = Integer.parseInt(st.nextToken());
